LIBBCM-38 Create a password reset form for Avalon
This adds the ability to send a password reset email for users that are
using Omniauth Identity to log-in. Users go to the login and can request
a password reset. If their password has a matching Identity record, a
reset URL with a token is sent to their address. When the click the
link, they are given a "Create a new password" form. After the form is
submitted, their password will be updated and the user will be logged
in.

https://issues.umd.edu/browse/LIBBCM-38