LIBAVALON-173. Specify Docker containers to run in "docker-compose"
Running `docker-compose up` without qualification runs all the Docker
containers in the "docker-compose.yml".

In the case of Avalon, running all the containers in not necessary,
because some of the containers in the "docker-compose.yml" file are
"test" containers used for testing (which require additional setup to
run properly).

While Avalon will still start properly, the running of the test
containers causes distracting warnings and errors to be emitted to the
console log.

The Avalon "README.md" indicates that the system should be started using
`docker-compose up avalon worker` which only runs the containers
necessary for Avalon (skipping the test containers). Modified the
"UMD-README.md" to use this command, to avoid the spurious
warnings/errors in the log.

https://issues.umd.edu/browse/LIBAVALON-173