LIBAVALON-106. Set default thumbnail for videos
Modifed MasteFile generation to replace the "thumbnail" image derived
from the video file, with a stock default thumbnail image.

This prevents a multitude of videos with black/empty thumbnail images
from being displayed on the "Browse" page, and provides a visual
indication to curators that a thumbnail needs to be chosen.

Added a "set_default_thumbnail" method the "MasterFile" class, and
modified the MasterFileBuilder (called when a item is added manually)
and the "Avalon::Batch::Entry" class (called when an item is added
via a batch process) to call the method.

https://issues.umd.edu/browse/LIBAVALON-106