LIBAVALON-178. Modify tests to not delete fixture file
The "master_files_controller_spec.rb" tests contained tests that
(due to the changes made in LIBAVALON-128) were deleting the
"video short.mp4" and "videoshort.mp4" fixture files.

Added the "with_temporary_copy_of_file" helper method so that the
tests could operate on a temporary copy of the fixture files, so
that the original file would not be changed.

This enabled a number of tests that were failing to pass, as the
fixture file they expect still present.

https://issues.umd.edu/browse/LIBAVALON-178