LIBAVALON-178. Mark all failing tests as "pending"
Tests in the "avalon-develop" branch are failing, likely due to changes
made to the upstream on which the tests were based.

We do not currently have the time/resources to immediately fix all the
tests, but would like to be alerted to any new failing tests via
Jenkins.

rspec has two different markers to exclude tests, "skip" and "pending".
A "skipped" test is simply not run, and the presence of the skipped test
is simply noted in the test report. There are currently 27 tests marked
as skipped from the upstream codebase.

A "pending" test is a test that still needs to be implemented or fixed.
rspec will still run the test, but its failure will not be marked as an
actual failure. Also, if a pending test is run an actually succeed, that
is indicated in the report as a failure (since the test is expected to
fail).

Marking tests as "pending" as it is expected that at least some of the
tests are failing because of "infrastructure" issues such as login
authorization. Using "pending" will make it simpler to re-enable the
tests when the appropriate mechanism is found to correct the
underlying infrastructure issue, as the pending tests will be marked as
"failed" (because they now pass, where they are expected to fail).

https://issues.umd.edu/browse/LIBAVALON-178