LIBAVALON-197. Add option to download the access copy
Added "master_file_download" permission to "ability.rb", which allows
a master file (i.e., "access copy") to be downloaded if the user is
a member of the collection the media_object associated with the master
file is part of. Due to the way permissions are setup, "admin" users
bypass this permission check entirely.

Updated "app/controllers/media_objects_controller.rb" to determine if
master file downloads are allowed. In the case of multiple master files,
all master files must be downloadable (this is not expected to be an
issue in practice, as all the files should be part of the same
collection).

Added a "Downloads" section to
"app/views/media_objects/_metadata_display.html.erb" that presents a
download link for each master file, if permitted.

Added "download" endpoint to the
"/app/controllers/master_files_controller.rb" that enables
a master file (i.e., "access copy") to be downloaded, if permitted.

Added the "download" endpoint to "config/routes.db"

Added RSpec tests for new functionality.

https://issues.umd.edu/browse/LIBAVALON-197