LIBAVALON-208. AccessToken expiration date can only be set on create
Modified the AccessToken model so that the expiration date cannot be
modified after the model is created.

This was done to simplify the handling of the "expired" field, which
is used by a cron job to determine which access tokens have passed their
expiration dates, and so need to have the read group removed from the
associated media object.

Not allowing the expiration date to be changed simplifies the model,
because it avoids the need to provide "unexpire" functionality. Such
"unexpire" functionality is not expected to be needed because a new
access token with the desired expiration date can be easily created.

https://issues.umd.edu/browse/LIBAVALON-208