LIBAVALON-213. Added a Download button to the media object detail page
Added a "Download" button next to the "Share" button on the media object
detail page.

Note: The "Download" button is placed to the left of the "Share" button,
(not to the right, as suggested in the issue description), because
left-clicking the "Share" button displays a tabbed panel. If the
"Download" button were placed to the right, the "Download" button
would be displaced to underneath the opened panel. It was not
immediately obvious how to correct this, so simply moved the
"Download" button to the left of the "Share" button.

Left-clicking "Download" button opens up a modal dialog, which shows a
bulleted list of the available downloads. This is to handle the
(potential) case of a media object have multiple master files to
download, and is similar in appearance to the "Downloads" panel in the
right sidebar.

The "Download" button functionality is provided in the
"app/views/media_objects/_download.html.erb" helper file, and is
configured to display only if there are one or more masterfiles, and
master file downloads are allowed.

The "app/views/media_objects/_item_view.html.erb" helper file was
modified to place the "Download" button only if master file downloads
are allowed.

Added CSS configuration for the "Download" button to the
"app/assets/stylesheets/avalon/_buttons.scss" file, following the
example of the "Share" button.

https://issues.umd.edu/browse/LIBAVALON-213