LIBAVALON-205. Initial "Access Tokens" panel implementation
Initial implementation of an "Access Tokens" panel on a media object's
"Access Control" page.

Added an "Access Tokens" panel to the the media object's
"Access Control" page that shows any active tokens, and a
"Create Access Token" button.

Modified the "new" method in the "AccessTokensController" class to
accept a "media_object_id" parameter, to use as the default for the
access token form.

https://issues.umd.edu/browse/LIBAVALON-205