LIBAVALON-205. Updates to AccessToken for invalid expiration dates
Updated AccessToken model and AccessTokensController to better handle
various invalid expiration dates. Some of these changes were
necessitated by the change in LIBAVALON-203, making it invalid for
an access token to be saved with an expiration date in the past.

Removed some tests that were predicated on an access token being
created with an expiration date in the past, as that should no longer
be possible.

https://issues.umd.edu/browse/LIBAVALON-205