LIBAVALON-216. Updated AccessToken model to allow create by depositors
Modified the "AccessToken" model validation logic use the "is_member_of"
(instead of "is_editor_of") method from the Ability object. This allows
depositors (as well as managers and editors) of the media object's
collection to create access tokens for the media object.

Also added a "media_object_id=" setter to ensure that the cached
"@media_object" instance variable will be updated if the
"media_object_id" value is changed.

https://issues.umd.edu/browse/LIBAVALON-216