LIBAVALON-216. Modified AccessToken factory to default to "admin" user
Modified the AccessToken factory to default to creating the access token
using a "admin" user. This was done because of changes to the
AccessToken model, which requires the user associated with the token to
be either an admin or a manager/editor/depositor of the media object's
collection.

This change restores the ability to create an AccessToken model
using a simple "FactoryBot.create(:access_token)", which is expected
by some tests.

https://issues.umd.edu/browse/LIBAVALON-216