LIBAVALON-226. Display "Manage Access Tokens" entry in navigation bar
Added "Manage Access Tokens" entry in the "Manage" dropdown on the
navigation bar. Placed the entry first, so that the items in the
dropdown would be in alphabetical order.

Restricted display of the "Manage Access Tokens" entry to users
that have the AccessToken "manage" ability, which should be
administrators (which can "manage" anything), or users that are
managers, editors, or depositors of any collection.

https://issues.umd.edu/browse/LIBAVALON-226