LIBAVALON-264. Modified controllers/views to support IP Manager failure
Modified the controllers to gracefully handle the inability to connect
to the IP Manager. The views were modified to replace the
"UMD IP Manager Group" section in the form with an error banner.

Note: When Avalon is unable to contact the IP Manager, all
IP Manager-based access by users will not work (i.e., users will not
be able to access Avalon items they would otherwise have access to).

Fixed the call to the UmdIPManager in "ability.rb" to call the
UmdIPManager method, instead of the API method directly.

Added test cases.

https://issues.umd.edu/browse/LIBAVALON-264