LIBAVALON-238. Updated Ability tests for streaming ability
Expanded the "spec/models/ability_spec.rb" tests for streaming to
cover unpublished/published media objects, and (for published media
objects) whether the media object's item access is private, restricted,
or public.

When an access token is provided, a media object is only streamable
if it is published. Having an access token does not provide streaming
access to an unpublished media object.

These tests attempt to codify the permissions outlined in
https://confluence.umd.edu/pages/viewpage.action?pageId=635241505

https://issues.umd.edu/browse/LIBAVALON-238