LIBAVALON-314. Corrected tests reliance on a particular URL host
These UMD-added tests were modified to be consistent with the URL
handling of the other stock Avalon tests in the same file, where
the URL path is checked, not the full URL.

The stock Avalon tests in this file were originally using the
full URL in their tests, but they were subsequently changed to rely
only on the URL path. These tests were added by UMD, so were not
changed as part of the Avalon 7.4 merge. Changing the UMD tests makes
them consistent with the other tests in this file.

https://umd-dit.atlassian.net/browse/LIBAVALON-314