LIBAVALON-314. Added "download" exception to "set_masterfile" action
The "set_masterfile" method and "before_action :set_masterfile"
directive were added in Avalon 7.3.

This addition was changing the way the UMD-added "download" method
was behaving when given an invalid masterfile id, because the invalid
check was being performed in the "set_masterfile" method, not in the
"download" method. This was causing the
"MasterFilesController#download fails if invalid id is provided" test in
"app/controllers/master_files_controller.rb" to fail.

The "download" method as originally written by UMD provides very
specific handling of an invalid masterfile id, displaying a specific
flash "alert" message, and redirecting the browser to either the
referring page, or the home page. The "set_masterfile" method, as
written by Avalon, sends a 404 response with a different flash "notice"
message.

Since "download" is a UMD-custom method, it seems best to leave it
behavior (both the alert message and redirect) in place. This was
accomplished by simply adding the "download" method to the list
of exceptions in the "before_action :set_masterfile" directive. This
decision can obviously be reassessed in the future.

Also added "UMD Customization" markers around the UMD changes in the
file, including an unrelated customization in the "get_frame" method.
These customization markers are intended to make it simpler to identify
UMD customizations when updating to future Avalon versions.

https://umd-dit.atlassian.net/browse/LIBAVALON-314