LIBAVALON-353. ruby-saml - Mitigate CVE-2024-45409
Added the "ruby-saml" version to the "Gemfile" and pinned to
1.12.3 (or later 1.12.x) version to mitigate CVE-2024-45409
(see https://github.com/advisories/GHSA-jw9c-mfg7-9rx2).

Without this pinning of the "ruby-saml" gem to 1.12.3, the "omniauth-saml"
gem will cause ruby-saml v1.14.0 to be used, which contains the
vulnerability.

It is currently not possible to update the "omniauth-saml" gem version
to the patched v2.2.1, because it requires Ruby 3.1 or greater.

This customization can likely be removed when upgrading to a later
ArchivesSpace version (as long as the "ruby-saml" gem matches a version
containing the fix).

https://umd-dit.atlassian.net/browse/LIBAVALON-353