LIBAVALON-362. Workaround to OmniAuth Devise SAML redirect loop
After upgrading to Avalon 7.6, left-clicking the "Sign In" button
caused all browsers to show an error page (for example, in Chrome:
"This page isn’t working - av-local redirected you too many times.").

This appears to be caused by the "Rails.application.reloader.to_prepare"
block that wraps the Devise configuration in the "devise.rb" file.

It is not quite clear whether commenting out the wrapper is the right
solution, but it does appear to resolve redirect loop.

Added a "Configuration Oddities" section to the "UMD-README.md" file
to provide information and additional resources for this change.

https://umd-dit.atlassian.net/browse/LIBAVALON-362