LIBAVALON-362. Set "Lax" SameSite attribute for session cookie
Setting the "SameSite" attribute of the Avalon session cookie to
"Lax" when running in the local development environment.

Without this setting, the Firefox browser rejects the session cookie,
with the following error message:

> Cookie “_session_id” rejected because it has the “SameSite=None”
> attribute but is missing the “secure” attribute.

and without the session cookie, Avalon goes into an infinite redirect
loop.

This occurs because when not set, Firefox defaults the "SameSite"
attribute to "None" (which requires the "Secure" flag, except when the
hostname is "localhost"), and the local development version of Avalon
has been customized to use "av-local" (instead of "localhost"), due to
UMD CAS restrictions.

Updated "UMD-README.md" with information about this configuration
customization.

https://umd-dit.atlassian.net/browse/LIBAVALON-362