LIBAVALON-362. Test fix - Skip invalid tests in api_authentication_spec
Marked tests as "pending" in "api_authentication_spec.rb" that are no
longer valid because of UMD customizations to the "custom_permissions"
method in "app/models/ability.rb".

Using "pending" instead of "skip" so RSpec will notify us if the tests
suddenly start passing again (in which case we can remove "pending")

For the "sets the session information properly" test, it appears that
the test relies on Avalon throwing a "CanCan::AccessDenied" exception
to prevent the user_session from being destroyed in the
"handle_api_request" method of
"app/controllers/application_controller.rb".

The other tests appear to rely on media objects being unauthorized
that are authorized under the UMD permission scheme.

https://umd-dit.atlassian.net/browse/LIBAVALON-362