LIBAVALON-362. Test fix - Skip failing test in capybara_homepage_spec
Marked test as "pending" in "capybara_homepage_spec.rb" because it does
not pass due to UMD customization of the "registrable" flag
in "avalon/config/settings.yml".

This was consistent with the other tests in this file. Updated the
"pending" comment in the other tests to be consistent (and to correct
"setting.yaml" to "setting.yml".

https://umd-dit.atlassian.net/browse/LIBAVALON-362