LIBAVALON-362. Update "authentication.rb" for "domain" Setting changes
In "config/settings.yml", the "domain" parameter was changed from
a String ("http://av-local:3000") to a Hash
{ domain: { host: av-local, port: 3000, protocol: http } }

Updated the "lib/avalon/authentication.rb" file to use the Hash
form instead of the String form when setting up the "saml_config" hash.

Did this because the String form seem to be obsolete (although still
supported), and the String form was causing some Avalon tests to fail.

Updated "environment" for "avalon" in docker-compose.yml to set the
"domain" using individual components, instead of as a single single,
to be compatible with the changes in the "authentication.rb".

https://umd-dit.atlassian.net/browse/LIBAVALON-362