LIBAVALON-373. Development Environment documentation updates
Updated Docker Development Environment documentation, moving the
"mc alias" command to before the "createbuckets" step, as that
seemed to be a more natural place to put it. Note that the "mc alias"
command appears to be idempotent, so it can be run multiple times
on a user's workstation without ill effect.

Simplified the creation of the "sample-data" administrative user to
a single Docker command.

Parameterized the uploading of files to minio, so that it will be
simpler to upload different sample datasets. Also used the "mc mirror"
command to upload everything *except* the XLSX manifest file (as it
has an "--exclude" flag that "mc cp" does not provide), and then
used "mc cp" to upload the manifest.

https://umd-dit.atlassian.net/browse/LIBAVALON-373