LIBAVALON-400. Fix "Not unique MediaObjectIndexingJob" in Solr reindex
In Kubernetes, the "rails avalon:reindex" would consistently output
"Not unique MediaObjectIndexingJob" for each MediaObject when
This error appears to prevent all the fields
(such as "has_transcripts_bsi") from being written to the Solr entry.

The “Not unique MediaObjectIndexingJob” message appears to be because of
the use of the “activejob-uniqueness” gem
(https://github.com/veeqo/activejob-uniqueness), which uses Redis to
create a “lock” that is not being properly removed in some cases. For
example, after the Solr reindex is complete, the locks remain in Redis
in “activejob_uniqueness:media_object_indexing_job” (the state of Redis
can be investigated using a Redis CLI client like the
“Redis for VS Code” extension -
https://redis.io/docs/latest/develop/tools/redis-for-vscode/).

Attempting to clear the locks by running `ActiveJob::Uniqueness.unlock!`
in the Rails console, or even wiping out the entire Redis database by
running the `rails avalon:wipeout:redis` Rake task did not fix the
issue – the locks were recreated each time time the Solr reindex ran.

This customization was suggested by a July 17, 2025 chat with
Chris Colvard in the “avalon” Slack channel. With this change, the
“Not unique MediaObjectIndexingJob” issue no longer occurs, and the
Solr entries for MediaObjects are fully populated.

https://umd-dit.atlassian.net/issues/LIBAVALON-400