LIBAVALON-397. Re-implemented "Request from Special Collections" for 7.8
Re-implemented the "Request from Special Collections" button on the
media object detail page as a React component to be compatible with
Avalon 7.8.

* Added “app/javascript/components/AeonRequestForm.jsx”

This React component (helpfully generated by TerpAI) replaces the
 “app/views/media_objects/_aeon_request_form.html.erb” file, and
actually displays the “Request from Special Collections” button which
submits a hidden form to Aeon, that pre-populates the Aeon form with the
values for the media object.

* Modified the “MediaObjectRamp” configuration setup in
“app/views/media_objects/_item_view.html.erb” file to add an
“aeon_request” JavaScript Map that contains the media object information
needed by the “AeonRequestForm.jsx” React component. The key-value
mapping was taken from
“app/views/media_objects/_aeon_request_form.html.erb”.
The “aeonUrl” key was added, as it seemed useful to have all the
potentially configurable parameters in one location.

* Modified the “app/javascript/components/MediaObjectRamp.jsx”
  component:
  * Imported the “AeonRequestForm” React component
  * Added the “aeon_request” map as a parameter to the “MediaObjectRamp”
    React component
  * Modified the HTML to insert the AeonRequestForm React component in
    the correct location, and provided it with the “aeon_request” map

* Deleted the UMD-added
  “app/views/media_objects/_aeon_request_form.html.erb” file, as it is
  no longer needed

Also added a “umd_docs/UmdCustomizations.md” document providing a
high-level overview of the customization, and added a link to the
document in “UMD-README.md”.

https://umd-dit.atlassian.net/browse/LIBAVALON-397