LIBAVALON-396. Updated "avalon-docker" Docker images for 7.8.0-umd-0
Updated the Docker images generated from the "umd-lib/avalon-docker"
repository to use the "7.8.0-umd-0" tagged versions.

Note: As described in "umd-lib/avalon-docker/UMD-README.md", the
Docker images from the "umd-lib/avalon-docker" repository would
normally be created and tagged as needed when the repository underwent
significant changes (such as a version upgrade), and not normally as
part of the a "umd-lib/avalon" release, unless changes had been made
to the "umd-lib/avalon-docker" repository immediately prior to the
release.

https://umd-dit.atlassian.net/browse/LIBAVALON-396