LIBAVALON-397. Fix "Invalid DOM property `class`" error in console
When displaying the media object detail page, the following error is
seen in the console:

```
Warning: Invalid DOM property `class`. Did you mean `className`?
```

This is caused by the use of the "class" attribute in:

```
<div class="request-special">
```

which was a UMD customization for the "Request from Special Collections"
button.

Changing the line to

```
<div className="request-special">
```

fixes the warning.

https://umd-dit.atlassian.net/browse/LIBAVALON-397