LIBAVALON-398. Added "Downloads" button to MediaObjectRamp component
Added a "Downloads" button to display between the "Create Timeline"
and "Share" buttons on the media object detail page.

Implemented as the "UmdMasterFileDownloadsButton" React component, which
is incorporated into the "MediaObjectRamp" component.

Removed the "_download.html.erb" file as it is no longer used.

Moved the CSS directives for the button from the stock Avalon
"_buttons.scss" file to the UMD "custom.css" file, as it is not
necessary to place the directives in the stock Avalon file.

https://umd-dit.atlassian.net/browse/LIBAVALON-398