LIBAVALON-401. Restore "Access Token" panel to "Access Control" page
Restored the UMD custom "Access Token" panel to the
"Edit Media Object > Access Control" for media objects.

This is largely the same code as was in the
"app/views/modules/_access_control.html.erb" file which was
deleted in Avalon 7.7.

The only real change was switching to using "card" instead of "panel"
CSS styling for the outer panel, to match the styling of the other
Avalon panels on the page.

https://umd-dit.atlassian.net/browse/LIBAVALON-401