LIBAVALON-416. Allow manifest retrieval when using access token
Enable the master file manifest information to be retrieved when a
user presents an access token with "streaming" permission.

The previous Avalon 7.4 implementation of Access Tokens apparently
did not need this change because its media player would "fall back" to
non-manifest URLs that was provided. The updated media player in Avalon
7.8 does not have that fallback functionality.

https://umd-dit.atlassian.net/browse/LIBAVALON-416