LIBAVALON-409. Disable Jenkins builds for Avalon
Disabled Jenkins builds for Avalon by renaming the "Jenkinsfile" to
"Jenkinsfile.disabled". This prevents the "umd-lib pipeline" GitHub
organization pipeline from running any builds.

The "Webhooks" in the GitHub repository have also been disabled so
that pull requests or new branches do not trigger Jenkins builds.

The Jenkins builds are disabled because the Avalon tests currently
take 6+ hours to complete, and consume enough resources to hamper
the operaton of the CI servers.

https://umd-dit.atlassian.net/browse/LIBAVALON-409