LIBAVALON-436. Make user_session optional for allow_view_in_repo
Some unit tests that use this partial do not set the user_session,
and so were failing with an `undefined method `[]' for nil:NilClass`
error.

Since user_session appears to be optional in other code, and the
"allow_view_in_repo" is intended to be true, except when an LTI callback
is set (see LIBAVALON-424), making user_session optional here as well.

https://umd-dit.atlassian.net/browse/LIBAVALON-436